fix crash when minEjectSize is too low
Fixes issue #1527.
Size of cell would become negative if the minEjectSize setting was too low, and when the square root of the negative number is calculated, the result is NaN.
The fix was to add a check to make sure that after removing the ejected mass the player cell is not below playerMinSize.